requirements.txt
asgiref==3.2.10
boto3==1.14.39
botocore==1.17.39
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
django-cors-headers==3.4.0
django-filter==2.3.0
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
httpie==2.2.0
idna==2.10
jmespath==0.10.0
Pillow==7.2.0
psycopg2==2.8.5
Pygments==2.6.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
